refactor: Move clean up fixture to each test file
